Move writeTmpCompose to specific file
