+++ src/tasks/buildAndUpload/writeTmpCompose.ts
@@ -20,20 +20,3 @@ export function writeTmpCompose({
   const buildCompose = updateComposeImageTags(compose, manifest);
   writeCompose(buildCompose, { dir: rootDir, composeFileName });
 }
-
-export function writeReleaseCompose({
-  compose,
-  composeFileName = defaultComposeFileName,
-  manifest,
-  releaseDir
-}: {
-  compose: Compose;
-  composeFileName?: string;
-  manifest: Manifest;
-  releaseDir: string;
-}): void {
-  const releaseCompose = updateComposeImageTags(compose, manifest, {
-    editExternalImages: true
-  });
-  writeCompose(releaseCompose, { dir: releaseDir, composeFileName });
-}
